Move authorization logic to controller

+++ backend/src/main/java/pt/ulisboa/tecnico/socialsoftware/tutor/clarification/ClarificationController.java
@@ -5,6 +5,8 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import pt.ulisboa.tecnico.socialsoftware.tutor.clarification.dto.ClarificationRequestAnswerDto;
+import pt.ulisboa.tecnico.socialsoftware.tutor.exceptions.ErrorMessage;
+import pt.ulisboa.tecnico.socialsoftware.tutor.exceptions.TutorException;
 import pt.ulisboa.tecnico.socialsoftware.tutor.user.User;
 
 import java.security.Principal;
@@ -20,6 +22,10 @@ public class ClarificationController {
     public ClarificationRequestAnswerDto submitAnswer(Principal principal, @PathVariable int requestId, @RequestBody String answer) {
         User user = (User) ((Authentication) principal).getPrincipal();
 
+        if (user == null) {
+            throw new TutorException(ErrorMessage.AUTHENTICATION_ERROR);
+        }
+
         return clarificationService.submitClarificationRequestAnswer(user, requestId, answer);
     }
 
@@ -27,6 +33,10 @@ public class ClarificationController {
     @PreAuthorize("hasRole('ROLE_TEACHER') and hasPermission(#requestId, 'CLARIFICATION.ACCESS')")
     public void deleteAnswer(Principal principal, @PathVariable int requestId) {
         User user = (User) ((Authentication) principal).getPrincipal();
+
+        if (user == null) {
+            throw new TutorException(ErrorMessage.AUTHENTICATION_ERROR);
+        }
 
         clarificationService.deleteClarificationRequestAnswer(user, requestId);
     }
